requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
-e hg+https://fleetpal:Eod0ei9z@hg.resolve.bg/resolve/common-helpers@default#egg=common_helpers
    # via -r requirements.in
aniso8601==7.0.0
    # via graphene
asgiref==3.3.4
    # via django
certifi==2021.5.30
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via requests
click==8.0.1
    # via
    #   pip-tools
    #   rq
cryptography==35.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-rest-auth[with_social]==2.1.7
    # via -r requirements.in
django-allauth==0.44.0
    # via dj-rest-auth
django-anymail[postmark]==8.4
    # via -r requirements.in
django-cors-headers==3.7.0
    # via -r requirements.in
django-crispy-forms==1.12.0
    # via -r requirements.in
django-elasticsearch-dsl==7.2.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-map-widgets==0.3.1
    # via -r requirements.in
django-phonenumber-field[phonenumbers]==5.2.0
    # via -r requirements.in
django-rest-framework==0.1.0
    # via -r requirements.in
django-rq-email-backend==0.1.3
    # via -r requirements.in
django-rq==2.4.1
    # via
    #   -r requirements.in
    #   django-rq-email-backend
django-waffle==2.2.1
    # via -r requirements.in
django==3.2.4
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-anymail
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-phonenumber-field
    #   django-rq
    #   django-rq-email-backend
    #   djangorestframework
    #   graphene-django
djangorestframework==3.12.4
    # via
    #   dj-rest-auth
    #   django-rest-framework
    #   drf-extensions
dnspython==1.16.0
    # via eventlet
drf-extensions==0.7.0
    # via -r requirements.in
elasticsearch-dsl==7.3.0
    # via django-elasticsearch-dsl
elasticsearch==7.13.1
    # via elasticsearch-dsl
eventlet==0.30.2
    # via -r requirements.in
flake8==3.9.2
    # via -r requirements.in
graphene-django==2.15.0
    # via -r requirements.in
graphene==2.1.8
    # via graphene-django
graphql-core==2.3.2
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
greenlet==1.1.2
    # via eventlet
gunicorn==20.1.0
    # via -r requirements.in
idna==2.10
    # via requests
ijson==3.1.4
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
oauthlib==3.1.1
    # via requests-oauthlib
pep517==0.10.0
    # via pip-tools
phonenumbers==8.12.26
    # via django-phonenumber-field
pillow==8.4.0
    # via reportlab
pip-tools==6.1.0
    # via -r requirements.in
promise==2.3
    # via
    #   graphene-django
    #   graphql-core
    #   graphql-relay
psycopg2-binary==2.9.1
    # via -r requirements.in
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.3.1
    # via flake8
pyjwt[crypto]==2.3.0
    # via django-allauth
python-dateutil==2.8.1
    # via elasticsearch-dsl
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via -r requirements.in
redis==3.5.3
    # via
    #   -r requirements.in
    #   django-rq
    #   rq
reportlab==3.6.2
    # via -r requirements.in
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
rq==1.9.0
    # via
    #   -r requirements.in
    #   django-rq
    #   django-rq-email-backend
rx==1.6.1
    # via graphql-core
sentry-sdk==1.1.0
    # via -r requirements.in
singledispatch==3.6.2
    # via graphene-django
six==1.16.0
    # via
    #   django-elasticsearch-dsl
    #   elasticsearch-dsl
    #   eventlet
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   promise
    #   python-dateutil
    #   singledispatch
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via graphene-django
toml==0.10.2
    # via pep517
urllib3==1.26.5
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
